fix envelope generator

+++ src/EGUISection.cpp
@@ -124,28 +124,28 @@ void EGUISection::sliderValueChanged( Slider *pSlider )
    {
       for( Sample *pSample : samples() )
       {
-         getENV( pSample )->setAttack( (float)m_psAttack->getValue() );
+         getENV( pSample )->setAttack( m_psAttack->getValue() );
       }
    } else
    if( pSlider == m_psDecay )
    {
       for( Sample *pSample : samples() )
       {
-         getENV( pSample )->setDecay( (float)m_psDecay->getValue() );
+         getENV( pSample )->setDecay( m_psDecay->getValue() );
       }
    } else
    if( pSlider == m_psSustain )
    {
       for( Sample *pSample : samples() )
       {
-         getENV( pSample )->setSustain( (float)m_psSustain->getValue() );
+         getENV( pSample )->setSustain( m_psSustain->getValue() );
       }
    } else
    if( pSlider == m_psRelease )
    {
       for( Sample *pSample : samples() )
       {
-         getENV( pSample )->setRelease( (float)m_psRelease->getValue() );
+         getENV( pSample )->setRelease( m_psRelease->getValue() );
       }
    }
 }
